chore: move ITERATE_KEY to operations
