+++ packages/runtime-core/__tests__/apiLifecycle.spec.ts
@@ -16,7 +16,7 @@ import {
   TrackOpTypes,
   onRenderTriggered
 } from '@vue/runtime-test'
-import { ITERATE_KEY, DebuggerEvent, TriggerOpTypes } from '@vue/reactivity'
+import { DebuggerEvent, TriggerOpTypes, ITERATE_KEY } from '@vue/reactivity'
 
 // reference: https://vue-composition-api-rfc.netlify.com/api.html#lifecycle-hooks
